Router elements apperaed
