+++ src/lib/components/Chats/Chats.js
@@ -1,0 +1,19 @@
+import React, { Component } from 'react';
+import Aux from '../../../hoc/Aux/Aux';
+
+
+class Chats extends Component {
+  constructor(props) {
+    super(props);
+  }
+
+  render() {
+    return (
+      <Aux>
+        <div></div>
+      </Aux>
+    );
+  }
+}
+
+export default Chats;
